fix(sheet): numfmt pattern update with cell type
