+++ packages/sheets/src/basics/__tests__/cell-type.spec.ts
@@ -1,0 +1,39 @@
+/**
+ * Copyright 2023-present DreamNum Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { describe, expect, it } from 'vitest';
+import { CellValueType } from '@univerjs/core';
+import { checkCellValueType } from '../cell-type';
+
+describe('test util cell type ', () => {
+    it('should be able to get the correct type of cell value from "checkCellValueType"', () => {
+        expect(checkCellValueType('string', CellValueType.BOOLEAN)).toBe(CellValueType.STRING);
+        expect(checkCellValueType('string', CellValueType.NUMBER)).toBe(CellValueType.STRING);
+        expect(checkCellValueType('string', CellValueType.STRING)).toBe(CellValueType.STRING);
+        expect(checkCellValueType('0', CellValueType.STRING)).toBe(CellValueType.NUMBER);
+        expect(checkCellValueType('true', CellValueType.STRING)).toBe(CellValueType.BOOLEAN);
+        expect(checkCellValueType('false', CellValueType.STRING)).toBe(CellValueType.BOOLEAN);
+
+        expect(checkCellValueType(123, CellValueType.STRING)).toBe(CellValueType.NUMBER);
+        expect(checkCellValueType(123, CellValueType.BOOLEAN)).toBe(CellValueType.NUMBER); // not a valid boolean number, casted to number
+        expect(checkCellValueType(123, CellValueType.NUMBER)).toBe(CellValueType.NUMBER);
+        expect(checkCellValueType(1, CellValueType.NUMBER)).toBe(CellValueType.NUMBER);
+        expect(checkCellValueType(0, CellValueType.NUMBER)).toBe(CellValueType.NUMBER);
+
+        expect(checkCellValueType(1, CellValueType.BOOLEAN)).toBe(CellValueType.BOOLEAN); // it is valid boolean value
+        expect(checkCellValueType(0, CellValueType.BOOLEAN)).toBe(CellValueType.BOOLEAN); // it is valid boolean value
+    });
+});
