+++ packages/sheets/src/basics/__tests__/cell-value.spec.ts
@@ -1,0 +1,32 @@
+/**
+ * Copyright 2023-present DreamNum Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { describe, expect, it } from 'vitest';
+import { extractBooleanValue } from '../cell-value';
+
+describe('test utils cell value', () => {
+    it('should be able to cast values that can be casted to boolean', () => {
+        expect(extractBooleanValue(1)).toBe(true);
+        expect(extractBooleanValue(0)).toBe(false);
+        expect(extractBooleanValue(-123)).toBe(null);
+        expect(extractBooleanValue('1')).toBe(true);
+        expect(extractBooleanValue('0')).toBe(false);
+        expect(extractBooleanValue('89757')).toBe(null);
+        expect(extractBooleanValue('true')).toBe(true);
+        expect(extractBooleanValue('false')).toBe(false);
+        expect(extractBooleanValue('Michael Jackson')).toBe(null);
+    });
+});
