+++ packages/sheets/src/services/numfmt/numfmt.service.ts
@@ -23,6 +23,8 @@ import {
     Range,
 } from '@univerjs/core';
 
+import { getCellTypeByPattern } from '../../basics/cell-type';
+import { getCellValue } from '../../basics/cell-value';
 import type { INumfmtService } from './type';
 
 export class NumfmtService extends Disposable implements INumfmtService {
@@ -110,6 +112,13 @@ export class NumfmtService extends Disposable implements INumfmtService {
                         const newStyle = { ...oldStyle, n: { pattern: value.pattern } };
                         const styleId = styles.setValue(newStyle);
                         cell.s = styleId;
+
+                        // Setting the text format for a cell will set the CellValueType to text
+                        const type = getCellTypeByPattern(cell, value.pattern);
+                        if (cell.v !== undefined) {
+                            cell.t = type;
+                            cell.v = getCellValue(type, cell);
+                        }
                     }
                 });
             });
